+++ packages/promo-pages/src/components/homepage/VideoAppsTitle.tsx
@@ -1,11 +1,10 @@
 import React from 'react';
-import styles from './remotionvideoappstitle.module.css';
 
 export const VideoAppsTitle: React.FC = () => {
 	return (
-		<div className={styles.container}>
-			<h2 className={styles.title}>
-				Build video <span className={styles.highlight}>apps</span>
+		<div className={'text-center'}>
+			<h2 className={'font-brand text-4xl'}>
+				Build video <span className={'font-brand'}>apps</span>
 			</h2>
 			<p>
 				Use our suite of tools to build apps that lets others create videos.
@@ -16,8 +15,8 @@ export const VideoAppsTitle: React.FC = () => {
 
 export const PricingTitle: React.FC = () => {
 	return (
-		<div className={styles.container}>
-			<h2 className={styles.title}>Pricing</h2>
+		<div className={'text-center'}>
+			<h2 className={'font-brand text-4xl'}>Pricing</h2>
 			<br />
 		</div>
 	);
@@ -25,8 +24,8 @@ export const PricingTitle: React.FC = () => {
 
 export const DemoTitle: React.FC = () => {
 	return (
-		<div className={styles.container}>
-			<h2 className={styles.title}>Demo</h2>
+		<div className={'text-center'}>
+			<h2 className={'font-brand text-4xl'}>Demo</h2>
 			<br />
 		</div>
 	);
