--- packages/promo-pages/src/components/homepage/editor.module.css
@@ -1,43 +1,0 @@
-.row {
-	display: flex;
-	flex-direction: row;
-	justify-content: flex-end;
-	text-align: right;
-	align-items: center;
-}
-
-.title {
-	font-size: 2.5em;
-	font-weight: 700;
-	line-height: 1.1;
-	text-align: right;
-}
-
-.mp4 {
-	color: var(--ifm-color-primary);
-}
-
-@keyframes opacity {
-	from {
-		opacity: 0;
-	}
-	to {
-		opacity: 1;
-	}
-}
-
-.video {
-	animation: opacity 0.6s;
-}
-
-@media (max-width: 900px) {
-	.row {
-		flex-direction: column-reverse;
-		text-align: left;
-		align-items: flex-start;
-		justify-content: flex-start;
-	}
-	.title {
-		text-align: left;
-	}
-}
