--- packages/promo-pages/src/components/homepage/realvideos.module.css
@@ -1,49 +1,0 @@
-.realvideos {
-	display: flex;
-	flex-direction: row;
-}
-
-.realtitle {
-	font-size: 2.5em;
-	font-weight: 700;
-	line-height: 1.1;
-}
-
-.realgradient {
-	color: var(--ifm-color-primary);
-}
-
-.realmp4container {
-	flex: 1;
-	justify-content: flex-end;
-	align-items: center;
-	display: flex;
-}
-
-@media screen and (max-width: 900px) {
-	.realvideos {
-		flex-direction: column-reverse;
-	}
-	.realmp4container {
-		justify-content: flex-start;
-	}
-	.icon {
-		animation: none;
-		margin-bottom: 30px;
-	}
-}
-
-@keyframes goup {
-	from {
-		opacity: 0;
-		transform: translateY(50px);
-	}
-	to {
-		opacity: 1;
-		transform: translateY(0px);
-	}
-}
-
-.icon {
-	animation: goup 0.6s;
-}
