--- packages/promo-pages/src/components/homepage/remotionvideoappstitle.module.css
@@ -1,13 +1,0 @@
-.container {
-	text-align: center;
-}
-
-.title {
-	font-size: 2.5em;
-	font-weight: 700;
-	line-height: 1.1;
-}
-
-.highlight {
-	color: var(--ifm-color-primary);
-}
